+++ src/tr2/game/collide.c
@@ -302,8 +302,8 @@ int32_t __cdecl Collide_CollideStaticObjects(
 
     Room_GetNearbyRooms(x, y, z, coll->radius + 50, height + 50, room_num);
 
-    for (int32_t i = 0; i < g_DrawRoomsCount; i++) {
-        const ROOM *const room = &g_Rooms[g_DrawRoomsArray[i]];
+    for (int32_t i = 0; i < g_RoomsToDrawCount; i++) {
+        const ROOM *const room = &g_Rooms[g_RoomsToDraw[i]];
 
         for (int32_t j = 0; j < room->num_static_meshes; j++) {
             const STATIC_MESH *const mesh = &room->static_meshes[j];