+++ src/tr2/game/room_draw.h
@@ -2,6 +2,7 @@
 
 #include "global/types.h"
 
+void Room_MarkToBeDrawn(int16_t room_num);
 void __cdecl Room_GetBounds(void);
 void __cdecl Room_SetBounds(
     const int16_t *obj_ptr, int32_t room_num, const ROOM *parent);