+++ src/tr2/global/vars.c
@@ -6,4 +6,7 @@ const char *g_TR2XVersion = "TR2X (non-Docker build)";
 
 GAME_FLOW_DIR g_GF_OverrideDir = (GAME_FLOW_DIR)-1;
 
+int16_t g_RoomsToDraw[MAX_ROOMS_TO_DRAW] = { 0 };
+int16_t g_RoomsToDrawCount = 0;
+
 SDL_Window *g_SDLWindow = NULL;