+++ src/tr2/global/vars.h
@@ -7,4 +7,7 @@
 extern const char *g_TR2XVersion;
 extern GAME_FLOW_DIR g_GF_OverrideDir;
 
+extern int16_t g_RoomsToDraw[MAX_ROOMS_TO_DRAW];
+extern int16_t g_RoomsToDrawCount;
+
 extern SDL_Window *g_SDLWindow;